Patch simple alias table algorithm
Patch simple alias table algorithm in a minimally disruptive way, with
an explicit index for the highest or lowest bin, and breaking away when
no such bin is found (aliasing is then complete). Note that the "i"
index (looping n-1) times is merely a counter, which makes the algorithm
difficult to follow.

The parent commit resolved the sample failure case (10, 10, 1, 1), but
this is coincidental and the code remains fragile, because there is no
distinction between failing to find a high (or low) bin, and exhausting
the corresponding loop. Therefore excessive aliasing might be performed,
despite all bins being already equalized. In the failure case this
triggers an aliasing of the last bin with itself.